requirements.txt
python-dotenv
psycopg2-binary
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
passlib[argon2]
python-jose